+++ fmt/src/lib.rs
