+++ serde/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "sval_serde"
+version = "0.1.0"
+authors = ["Ashley Mannix <ashleymannix@live.com.au>"]
+edition = "2018"
+documentation = "https://docs.rs/sval_serde"
+description = "serde integration with sval"
+repository = "https://github.com/sval-rs/sval"
+license = "Apache-2.0 OR MIT"
+keywords = ["serialization", "no_std"]
+categories = ["encoding", "no-std"]
+publish = false
