+++ serde/tests/no_alloc/lib.rs
