+++ src/owned/tag.rs
@@ -1,0 +1,6 @@
+pub struct OwnedTag(Inner);
+
+enum Inner {
+    Owned(String),
+    Static(&'static str),
+}
