+++ src/stream/source.rs
